requirements.txt
attrs
aws-cdk.assets
aws-cdk.aws-apigateway
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling
aws-cdk.aws-autoscaling-common
aws-cdk.aws-autoscaling-hooktargets
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudwatch
aws-cdk.aws-codebuild
aws-cdk.aws-codecommit
aws-cdk.aws-codedeploy
aws-cdk.aws-codeguruprofiler
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codestarnotifications
aws-cdk.aws-cognito
aws-cdk.aws-ec2
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-ecs
aws-cdk.aws-efs
aws-cdk.aws-elasticloadbalancing
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-globalaccelerator
aws-cdk.aws-iam
aws-cdk.aws-kinesis
aws-cdk.aws-kinesisfirehose
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-lambda-destinations
aws-cdk.aws-logs
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-sam
aws-cdk.aws-secretsmanager
aws-cdk.aws-servicediscovery
aws-cdk.aws-signer
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.aws-stepfunctions
aws_cdk.aws_lambda_python
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.pipelines
aws-cdk.region-info
boto3
botocore
cattrs
certifi
chardet
charset-normalizer
constructs==3.3.69
idna
jmespath
jsii
numpy
pandas
publication
python-dateutil
pytz
requests
s3transfer
six
typing-extensions
urllib3
